requirements.txt
smolagents
huggingface_hub
langfuse
openinference
python-dotenv
